Add `order_valid_cs_pairs`.
Not the world's best name, but I can revisit.  Regardless, this sets in
motion the plan, which is to silently drop invalid CSIDs and hash up the
rest.

Truly interested parties can look at the map to see got through and
contributed.

Also applies a convenience conversion of "cs"- IDs.  Not so hep on the
magic, but meh.