Allow for generating with intermediates.
This will make it easier for doing handshakes(!) in a little bit.

We can either work from a whole bunch of intermediates or a combination
of CSKs and intermediates.

The maps in and out are slightly more expensive than they ought to be,
but it improves readbaility.  Can reconsider if it becomes an issue.